requirements.txt
# GUI
PySide6==6.5.2

# OCR y procesamiento de imágenes
paddleocr==2.6.1.3
paddlepaddle==2.4.2
Pillow==9.4.0

# Integración con OpenAI
openai==1.3.9

# Utilidades
pyperclip==1.8.2
pywin32==305
keyboard==0.13.5
pynput==1.7.6
requests==2.28.2
numpy==1.24.2
python-dotenv==1.0.0
cryptography==41.0.1